merge-ort: add comment to avoid surprise with new sub_flag variable
Commit 4057523a40 ("submodule merge: update conflict error message",
2022-08-04) added a sub_flag variable that is used to store a value from
enum conflict_and_info_types, but initializes it with an invalid value
of -1.  The code may never set it to a valid value, and use the invalid
one.  This can be surprising when reading over the code at first, but it
was intentional.  Add a comment making it clear that it is okay to be
using an invalid value, due to how it is used later.

Signed-off-by: Elijah Newren <newren@gmail.com>