merge-tree: show the original file names in the conflict output
When showing the conflict output in an `ls-files -u`-style form, it
makes sense to show conflicted files' names in the form that they were
written on disk (for example, in case of a file/directory conflict, we
would have written the file under a different name by appending a
`~<branch>` suffix).

However, when the intention is to perform a worktree-less merge, this
makes less sense, and causes more confusion than is necessary. Let's
show the original file names instead.

This patch also fixes a bug where the output of conflicts involving a
rename did not always show the correct file name for the respective
stages.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>