merge-tree -z: always show the original file name first
When printing the messages in a machine-readable format (i.e. in `-z`
mode), the intention is to support third-party software that wants to
present the conflicts in a "pretty" way to human readers.

To that end, `git merge-tree -z` prefixes the conflict messages by a
variable-size list of paths and a conflict identifier. This list and
this identifier are intended to be machine-parseable, the conflict
message is sort of free-form and not intended to be parsed.

Keeping in mind that the intended use case is to have third-party
software use `git merge-tree` to perform worktree-less merges and then
present the conflicts (if any) to human readers, it makes much more
sense to show the original file names of the involved files instead of
the ones we internally munged to allow for creating a tree object that
contains entries for both sides of the conflict (which requires them to
have different names).

So let's mention the original file names prominently, as first item in
that variable-size list of paths.

Note: For the modify/delete conflict type, we used to mention _only_ the
munged name in that path list. To allow for tools to read the tree
object produced by `git merge-tree -z` into a Git index in order to
resolve the conflicts, it is necessary to list not only the original
name but _also_ the munged name so that the item with the munged file
name can be removed from that Git index. Therefore, this patch teaches
`git merge-tree` to show both the original _and_ the munged name in that
list.

Also note: This patch changes the output of the remerge diff slightly:
whereas before, we printed the notice about a file/directory conflict
under the diff header mentioning the munged file name, we now print it
under a separate diff header that mentions the original file name.
That is the explanation why t4301 is touched by this patch.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Junio C Hamano <gitster@pobox.com>