merge: avoid searching for strategies with fewer than 0 conflicts
builtin/merge.c has a loop over the specified strategies, where if
they all fail with conflicts, it picks the one with the least number
of conflicts.

In the codepath that finds a successful merge, if an automatic commit
was wanted, the code breaks out of the above loop, which makes sense.
However, if the user requested there be no automatic commit, the loop
would continue looking for a "better" strategy.  Since it had just
found a strategy with 0 conflicts though, and it is not possible to
have fewer than 0 conflicts, the continuing search is guaranteed to be
futile.

While searching additional strategies won't cause problems other than
wasting energy, it is wasteful.  Avoid searching for other strategies
with fewer than 0 conflicts.

Signed-off-by: Elijah Newren <newren@gmail.com>