bisect--helper: migrate to OPT_SUBCOMMAND()
We just introduced a new way to process subcommands, and the
`bisect--helper` learns about it in preparation for replacing the
`git-bisect.sh` script altogether.

As a consequence, the usage strings are copied over from the scripted
version of the `git bisect` command. To avoid regressing on 959d670d1a4
(i18n: remove from i18n strings that do not hold translatable parts,
2022-01-31), we specifically do not enclose usage strings in `N_(...)`
that do not contain any translatable parts.

Note that a couple of `bisect_*()` functions are not converted into
`cmd_bisect_*()` functions directly, as they have callers other than the
`OPT_SUBCOMMAND()` one (and the original functions did not expect
a subcommand name to be passed as `argv[0]`, unlike the convention for
the `cmd_*()` functions. In those cases, we introduce wrapper functions
`cmd_*()` that also call the original function.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>