fuzz: reorganise the path for existing oss-fuzz fuzzers
This patch is aimed to provide a better organisation for oss-fuzz
fuzzers and to avoid top-level clustters in the git repository when
more fuzzers are introduced. In addition, grouping the fuzzers into
their own directory separate their application on fuzz-testing from
the core functionalities of the git code. It also provide better
and tidier structure the oss-fuzz fuzzing library to manage, locate,
build and execute those fuzzers for fuzz-testing purposes in future
development.

A new direrctory oss-fuzz has been created and existing fuzzers are
moved into the new directory. Makefile and .gitignore have been
fixed accordingly.

CC: Josh Steadmon <steadmon@google.com>
CC: David Korczynski <david@adalogics.com>
Signed-off-by: Arthur Chan <arthur.chan@adalogics.com>